Delelte netrw config file
I will stick to NERDTree for a while. It may not be strictly required as
netrw may be enough to do most of the stuff I use NERDTree for, but
still NERDTree is a very common setup and I can get more help for it
from the web.

Check out https://shapeshed.com/vim-netrw/ for more information about a
NERDTree free setup.