Tweak nerdtree
NerdTree is automatically open when a directory is passed as an
argument. This allows to open a project dir  with a nerdtree window and
an empty buffer and immediately serch for files with telescope. Without
the empty buffer (just the tree content displayed as buffer) triggering
Telescope find_files was leading to some strange behaviour.

I also removed the autocmd to prevent other buffer to replace nerdtree
because it was doing exactly the opposite when opening with a dir as
argument (the first file would always replace nerdtree split)