Put signature help below regular autocomplete
Most of the time signaure help is covering the autocomplete I need to
see. I'm not sure this will be enough though. Even if ray-x/lsp_signature.nvim
if really helpful for signature help its visual output in the wrong
place most of the time. Maybe I'll try lsp saga in the future.