Compact file list visualization for telescop
Add a file_only_entry_maker that renders fiename:linenumber. Such
visualization if helpful while looking for references or implementations
with lsp, where you only care for the file name (you can get the other
info from the preview).

It could be useful in other visualization as well(every time we usa a
preview?).
On second thought we almost replicated what telescope is already doing
for buffers picker. Maybe we better reuse that piece of code?