Speed up TestNearest mapping
Since every test related mapping starts with a `t` having a single `t`
for testing nearest test is slow because the command is executed after
the timeout (vim is waiting for another char).
Having it defined as `tt` removes the ambiguity and speed up execution,
even if a second char is required. The same result could be achieved
typing <leader>t<CR> but I figure just `tt` would make more sense to me.